# Minor performance improvements for analysis guards