## Description of the issue/feature this PR addresses

These improvements make the guards (especially `guard_retract`, `guard_submit` and `guard_reject`) to work slightly faster.

## Current behavior before PR

In a Worksheet with 23 analyses selected (6 submitted and 17 without being submitted) that are strongly dependent amongst them with calculations, the result of `ajax_transition` is ~1.249s:

```
2019-03-16 01:32:31 INFO senaite.core Execution of 'ajax_transitions' took 1.269209s
2019-03-16 01:32:34 INFO senaite.core Execution of 'ajax_transitions' took 1.261643s
2019-03-16 01:32:36 INFO senaite.core Execution of 'ajax_transitions' took 1.197963s
2019-03-16 01:32:39 INFO senaite.core Execution of 'ajax_transitions' took 1.224438s
2019-03-16 01:32:44 INFO senaite.core Execution of 'ajax_transitions' took 1.316256s
2019-03-16 01:32:46 INFO senaite.core Execution of 'ajax_transitions' took 1.222635s
```

## Desired behavior after PR is merged

In a Worksheet with 23 analyses selected (6 submitted and 17 without being submitted) that are strongly dependent amongst them with calculations, the result of `ajax_transition` is ~1.015s:

```
2019-03-16 01:30:47 INFO senaite.core Execution of 'ajax_transitions' took 1.009090s
2019-03-16 01:30:51 INFO senaite.core Execution of 'ajax_transitions' took 0.959412s
2019-03-16 01:30:53 INFO senaite.core Execution of 'ajax_transitions' took 1.081447s
2019-03-16 01:30:55 INFO senaite.core Execution of 'ajax_transitions' took 0.988112s
2019-03-16 01:30:57 INFO senaite.core Execution of 'ajax_transitions' took 0.980296s
2019-03-16 01:31:00 INFO senaite.core Execution of 'ajax_transitions' took 1.070923s
```

--
I confirm I have tested this PR thoroughly and coded it according to [PEP8][1]
and [Plone's Python styleguide][2] standards.

[1]: https://www.python.org/dev/peps/pep-0008
[2]: https://docs.plone.org/develop/styleguide/python.html